# go/model_message.go
/*
 * message exchange
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Message struct {

	Id int32 `json:"id"`

	Body string `json:"body"`

	Timer int32 `json:"timer"`

	Encrypted bool `json:"encrypted,omitempty"`

	Status string `json:"status,omitempty"`
}